code: moved admin locations space map into seperate component
