+++ src/components/admin-locations-space-map/README.md
@@ -1,0 +1,4 @@
+# AdminLocationsSpaceMap
+
+Renders a map that displays the current location of a space. This is used to both display the space
+and also modify the space's position.
